build: permit building without LLVM
This enables XCTest to be built without LLVM which is needed for
testing.  If `-DENABLE_TESTING=YES` is specified, either llvm must be
installed on the host, `-DLLVM_DIR` must be defined, or
`-DLLVM_MAIN_SRC_DIR` must be defined.  This allows the tests to be run
without building LLVM.  Default the tests to enabled to preserve
existing behaviour.